requirements.txt
python-binance
selenium
pandas
dash
plotly